feat: move hooks to services (#1189)
* feat: when user disconnects, formData is removed

* fix: warning react for uncontrolled to controlled component

* refactor: remove hooks dir and move hooks to services

* fix: relative import to resolve vite problems with alias
