+++ packages/app/src/components/ClientAuthenticatedOnly.tsx
@@ -1,7 +1,8 @@
 import type { PropsWithChildren } from "react";
 import React from "react";
 import { useMountedState } from "react-use";
-import { useUser } from "../hooks/useUser";
+
+import { useUser } from "@services/apiClient";
 
 export const ClientAuthenticatedOnly = ({ children, ...delegated }: PropsWithChildren<unknown>) => {
   const isMounted = useMountedState();
