+++ packages/app/src/design-system/base/HeaderBody.tsx
@@ -1,7 +1,6 @@
 import Link from "next/link";
 
-import { useUser } from "../../hooks/useUser";
-import { useFormManager } from "../../services/apiClient/form-manager";
+import { useFormManager, useUser } from "../../services/apiClient";
 import { FormButton } from "./FormButton";
 import { Logo } from "./Logo";
