--- packages/app/src/hooks/index.ts
@@ -1,4 +1,0 @@
-export * from "./useCallbackOnMount";
-export * from "./useConfig";
-export * from "./useSearch";
-export * from "./useStats";
