--- packages/app/src/hooks/useCallbackOnMount.ts
@@ -1,9 +1,0 @@
-import { useCallback } from "react";
-import type { AnyFunction } from "@common/utils/types";
-
-export const useCallbackOnMount = (callback: AnyFunction) => {
-  return useCallback(() => {
-    callback();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-};
