+++ packages/app/src/pages/ecart-rep/declarant.tsx
@@ -4,8 +4,6 @@ import { useCallback, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { useUser } from "../../hooks/useUser";
-import { useFormManager } from "../../services/apiClient/form-manager";
 import type { NextPageWithLayout } from "../_app";
 import { ClientAuthenticatedOnly } from "@components/ClientAuthenticatedOnly";
 import { RepartitionEquilibreeLayout } from "@components/layouts/RepartitionEquilibreeLayout";
@@ -19,6 +17,7 @@ import {
   FormLayout,
   FormLayoutButtonGroup,
 } from "@design-system";
+import { useFormManager, useUser } from "@services/apiClient";
 
 const title = "Informations déclarant";
