+++ packages/app/src/pages/ecart-rep/email.tsx
@@ -4,8 +4,6 @@ import React, { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { useUser } from "../../hooks/useUser";
-import { requestEmailForToken } from "../../services/apiClient/token";
 import type { NextPageWithLayout } from "../_app";
 import type { FeatureStatus } from "@common/utils/feature";
 import { ClientOnly } from "@components/ClientOnly";
@@ -21,6 +19,7 @@ import {
   FormLayout,
   FormLayoutButtonGroup,
 } from "@design-system";
+import { requestEmailForToken, useUser } from "@services/apiClient";
 
 const title = "Validation de l'email";
