+++ packages/app/src/pages/ecart-rep/entreprise.tsx
@@ -1,11 +1,9 @@
 import { useRouter } from "next/router";
 
-import { useConfig } from "../../hooks";
-import { useUser } from "../../hooks/useUser";
-import { useFormManager } from "../../services/apiClient/form-manager";
 import type { NextPageWithLayout } from "../_app";
 import { RepartitionEquilibreeLayout } from "@components/layouts/RepartitionEquilibreeLayout";
 import { FormButton, FormGroup, FormInput, FormGroupLabel, FormLayout, FormLayoutButtonGroup } from "@design-system";
+import { useConfig, useFormManager, useUser } from "@services/apiClient";
 
 const title = "Informations entreprise";
