+++ packages/app/src/pages/ecart-rep/periode-reference.tsx
@@ -6,8 +6,6 @@ import React, { useCallback, useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { useUser } from "../../hooks/useUser";
-import { useFormManager } from "../../services/apiClient/form-manager";
 import type { NextPageWithLayout } from "../_app";
 import { ClientAuthenticatedOnly } from "@components/ClientAuthenticatedOnly";
 import { RepartitionEquilibreeLayout } from "@components/layouts/RepartitionEquilibreeLayout";
@@ -21,6 +19,7 @@ import {
   FormLayout,
   FormLayoutButtonGroup,
 } from "@design-system";
+import { useFormManager, useUser } from "@services/apiClient";
 
 const title = "Période de référence";
