+++ packages/app/src/services/apiClient/token.ts
@@ -1,4 +1,4 @@
-import { fetcher } from "@common/utils/fetcher";
+import { fetcher } from "./fetcher";
 
 export const requestEmailForToken = (email: string, url: string) =>
   fetcher("/token", {
