+++ packages/app/src/services/apiClient/useMe.ts
@@ -1,5 +1,6 @@
 import useSWR from "swr";
-import { fetcher } from "../common/utils/fetcher";
+
+import { fetcher } from "./fetcher";
 
 // TODO: type à confirmer par rapport au endpoint /!\
 type DeclarationSummary = {
