+++ packages/app/src/services/apiClient/useSearch.ts
@@ -2,11 +2,11 @@ import moize from "moize";
 import useSWRInfinite from "swr/infinite";
 
 import type { SWRInfiniteKeyLoader } from "swr/infinite";
-import type { FetcherInfiniteReturn } from "../common/utils/fetcher";
+import type { FetcherInfiniteReturn } from "./fetcher";
 
-import { fetcher } from "../common/utils/fetcher";
+import { fetcher } from "./fetcher";
 import type { CompaniesType, CompanyType } from "@common/models/company";
-import { buildUrlParams } from "@common/utils/url";
+import { buildUrlParams } from "../../common/utils/url";
 
 export type SearchCompanyParams = {
   departement?: string;
