+++ packages/app/src/services/apiClient/useUser.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect } from "react";
 import create from "zustand";
 import { persist } from "zustand/middleware";
 
-import { useFormManager } from "../services/apiClient/form-manager";
+import { useFormManager } from "./useFormManager";
 import { useMe } from "./useMe";
 
 type UserStore = {
