+++ packages/app/tsconfig.json
@@ -11,6 +11,9 @@
       "@components/*": [
         "packages/app/src/components/*"
       ],
+      "@services/*": [
+        "packages/app/src/services/*"
+      ],
       "@design-system": [
         "packages/app/src/design-system/index"
       ]
